+++ libraries/base/changelog.md
@@ -25,6 +25,11 @@
     `newAlignedPinnedByteArray#` actually always assumed this but we didn't
     document this fact in the user facing API until now.
 
+    `Generic1`, and `Data` instances for `GHC.Tuple.Solo`.
+
+  * Under POSIX, `System.IO.openFile` will no longer leak a file descriptor if it
+    is interrupted by an asynchronous exception (#19114, #19115).
+
 ## 4.15.0.0 *TBA*
 
   * `openFile` now calls the `open` system call with an `interruptible` FFI